pyproject.toml
[project]
name = "tjm-project"
version = "0.1.0"
description = "Vision-based desktop automation with dynamic icon grounding"
requires-python = ">=3.10"
dependencies = [
    "pyautogui>=0.9.54",
    "Pillow>=10.0.0",
    "requests>=2.31.0",
    "pygetwindow>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
